+++ cmd/nvidia-cdi-hook/cudacompat/cuda-elf-header_test.go
@@ -33,12 +33,13 @@ func TestGetCUDACompatElfHeader(t *testing.T) {
 	dataRoot := filepath.Join(moduleRoot, "testdata", "compat")
 
 	testCases := []struct {
-		description string
-		filename    string
-		expected    *compatElfHeader
+		description   string
+		filename      string
+		expected      *compatElfHeader
+		expectedError string
 	}{
 		{
-			description: "wip",
+			description: "libcuda.so.575.57.08",
 			filename:    "libcuda.so.575.57.08",
 			expected: &compatElfHeader{
 				Format:      1,
@@ -48,13 +49,26 @@ func TestGetCUDACompatElfHeader(t *testing.T) {
 			},
 		},
 		{
-			description: "wip",
+			description: "libcuda.so.590.44.01",
 			filename:    "libcuda.so.590.44.01",
 			expected: &compatElfHeader{
 				Format:      1,
 				CUDAVersion: "13.1",
 				Driver:      []int{535, 550, 570, 575, 580, 590},
 				Device:      []int{1, 2, 7, 8, 9, 10, 11, 12, 13, 14},
+			},
+		},
+		{
+			description:   "invalid json",
+			filename:      "libcuda.invalid.json.so.99.88",
+			expectedError: "could not unmarshal JSON data",
+		},
+		{
+			description: "orin-13.2.1",
+			filename:    "libcuda.orin.13.2.1.so.1.1",
+			expected: &compatElfHeader{
+				Format:      1,
+				CUDAVersion: "13.2",
 			},
 		},
 	}
@@ -64,6 +78,11 @@ func TestGetCUDACompatElfHeader(t *testing.T) {
 			libpath := filepath.Join(dataRoot, tc.filename)
 
 			h, err := GetCUDACompatElfHeader(libpath)
+			if tc.expectedError != "" {
+				require.ErrorContains(t, err, tc.expectedError)
+				require.Nil(t, h)
+				return
+			}
 			require.NoError(t, err)
 
 			require.EqualValues(t, tc.expected, h)